requirements.txt
requests~=2.32.4
validators~=0.35.0
psycopg2-binary~=2.9.10
pytest~=8.4.1